feat: TeaVM compilation of MC 26.1.2 classes working
- Downloaded MC 26.1.2 client JAR and 27 dependency libraries
- Created SLF4J browser-compatible stubs (Logger, LoggerFactory, Marker, etc.)
- Created TeaVM classlib patches for StackWalker, CompletableFuture, etc.
- Configured build with JDK 25 for Java 25 class file support
- Removed conflicting NIO buffer stubs (TeaVM classlib provides these)
- Fixed LWJGL stub compilation errors (GL11, GL15, GLFW)
- TeaVM now successfully compiles 279 source files including 39 MC/Mojang classes
- Updated CI/CD to auto-download MC JAR and libs from Mojang API
- classes.js grew from 180KB to 293KB with MC class references
- Build: BUILD SUCCESSFUL - TeaVM generates JavaScript output

+++ download/eaglercraft-26.1.2/sources/teavm/java/org/lwjgl/opengl/GL15.java
@@ -155,7 +155,7 @@ public class GL15 {
         public static void glBufferData(int target, byte[] data, int usage) {
                 Uint8Array arr = Uint8Array.create(data.length);
                 for (int i = 0; i < data.length; i++) {
-                        arr.set(i, data[i] & 0xFF);
+                        arr.set(i, (short)(data[i] & 0xFF));
                 }
                 PlatformOpenGL._wglBufferData(target, arr, usage);
         }
@@ -197,7 +197,7 @@ public class GL15 {
         public static void glBufferSubData(int target, int offset, byte[] data) {
                 Uint8Array arr = Uint8Array.create(data.length);
                 for (int i = 0; i < data.length; i++) {
-                        arr.set(i, data[i] & 0xFF);
+                        arr.set(i, (short)(data[i] & 0xFF));
                 }
                 PlatformOpenGL._wglBufferSubData(target, offset, arr);
         }
